configure H2 in-memory database + scheme and data scripts

+++ src/main/java/com/twelveweeks/domain/transactions/Finance.java
@@ -14,8 +14,8 @@ public interface Finance {
     Date getDate();
     void setDate(Date date);
 
-    Long getUserId();
-    void setUserId(Long userId);
+    Integer getUserId();
+    void setUserId(Integer userId);
 
     Category getCategory();
     void setCategory(Category category);
